fix(memprovider): use int64 type in IntEvaluation to match API
The IntEvaluation method was using genericResolve[int] which caused
TYPE_MISMATCH errors when users stored int64 values in Variants.

The FeatureProvider interface defines IntEvaluation with int64:
  IntEvaluation(ctx, flag string, defaultValue int64, ...) IntResolutionDetail

The type assertion in genericResolve should use int64 to match.

Before this fix:
- User stores: Variants{"max": int64(100)}
- genericResolve[int] tries: value.(int)
- Type assertion fails because int64 != int
- Returns default value with TYPE_MISMATCH error

After this fix:
- User stores: Variants{"max": int64(100)}
- genericResolve[int64] tries: value.(int64)
- Type assertion succeeds
- Returns correct value

Updated tests to use explicit int64 types in Variants to match the
expected usage pattern for the IntEvaluation API.

+++ openfeature/memprovider/in_memory_provider.go
@@ -95,10 +95,10 @@ func (i InMemoryProvider) IntEvaluation(ctx context.Context, flag string, defaul
 	}
 
 	resolveFlag, detail := memoryFlag.Resolve(defaultValue, flatCtx)
-	result := genericResolve[int](resolveFlag, int(defaultValue), &detail)
+	result := genericResolve[int64](resolveFlag, defaultValue, &detail)
 
 	return openfeature.IntResolutionDetail{
-		Value:                    int64(result),
+		Value:                    result,
 		ProviderResolutionDetail: detail,
 	}
 }
